Add support for persisting cloudcmd config and https
This commit adds the ability to persist cloudcmd configuration
to a file (cloudcmd.json) on an external volume. This allows config changes
to be preserved even after Docker containers are recycled. This commit also
adds support for enabling https on the cloudcmd web server. https is the
default but http can be configured through a 'cloudcmd-https.json' config file.